Fix BE and iframe builds, but breaks SDK build.
