+++ src/api/modules/dispatch/dispatch.foreground.ts
@@ -4,7 +4,7 @@ import type Transaction from "arweave/web/lib/transaction";
 import type { TransformFinalizer } from "~api/foreground/foreground-modules";
 import type { ModuleFunction } from "~api/module";
 import type { DispatchResult } from "./index";
-import { sendChunk } from "../sign/chunks";
+import { sendChunk } from "~isomorphic-chunking";
 
 const foreground: ModuleFunction<Record<any, any>> = async (
   transaction: Transaction
