+++ src/api/modules/sign/chunks.ts
@@ -1,8 +1,4 @@
 import type { Tag } from "arweave/web/lib/transaction";
-import type { ApiCall, ApiResponse } from "shim";
-import { nanoid } from "nanoid";
-import { isApiErrorResponse } from "~utils/messaging/common/messaging.utils";
-import { isomorphicSendMessage } from "~utils/messaging/messaging.utils";
 
 /**
  * The chunk of the transaction signing
@@ -26,75 +22,6 @@ const chunks: {
   origin: string; // tabID for verification
   rawChunks: Chunk[]; // raw chunks to be reconstructed
 }[] = [];
-
-/**
- * Send a chunk to the background script
- *
- * @param chunk Chunk to send
- *
- * @returns Response from the background
- */
-export const sendChunk = (chunk: Chunk) =>
-  new Promise<void>(async (resolve, reject) => {
-    const callID = nanoid();
-
-    // construct message
-    const message: ApiCall = {
-      app:
-        import.meta.env?.VITE_IS_EMBEDDED_APP === "1"
-          ? "wanderEmbedded"
-          : "wander",
-      // TODO: Add Wallet API version:
-      version: "",
-      callID,
-      type: "chunk",
-      data: chunk
-    };
-
-    if (import.meta.env?.VITE_IS_EMBEDDED_APP === "1") {
-      const res = await isomorphicSendMessage({
-        destination: "background",
-        messageId: "chunk",
-        data: message
-      });
-
-      if (isApiErrorResponse(res)) {
-        return reject(res.data);
-      }
-
-      resolve();
-    } else {
-      // send message
-      window.postMessage(message, window.location.origin);
-
-      // wait for the background to accept the chunk
-      window.addEventListener("message", callback);
-
-      // callback for the message
-      function callback(e: MessageEvent<ApiResponse<number>>) {
-        const { data: res } = e;
-
-        if (!res) return;
-
-        // returned chunk index
-        const index = res.data;
-
-        // ensure we are getting the result of the chunk sent
-        // in this instance / call of the function
-        if (res.callID !== callID) return;
-
-        // check for errors in the background
-        if (isApiErrorResponse(res) || typeof index === "string") {
-          reject(res.data);
-        } else {
-          resolve();
-        }
-
-        // remove listener
-        window.removeEventListener("message", callback);
-      }
-    }
-  });
 
 /**
  * Handle incoming chunks and add them to the chunk storage
