+++ src/api/modules/sign/sign.foreground.ts
@@ -2,7 +2,7 @@ import type { SignatureOptions } from "arweave/web/lib/crypto/crypto-interface";
 import type { TransformFinalizer } from "~api/foreground/foreground-modules";
 import { createCoinWithAnimation } from "./animation";
 import type { ModuleFunction } from "~api/module";
-import { sendChunk } from "./chunks";
+import { sendChunk } from "~isomorphic-chunking";
 import {
   deconstructTransaction,
   type SplitTransaction
