+++ src/utils/messaging/strategies/iframe/iframe-messaging.strategy.ts
@@ -135,7 +135,7 @@ function getPostMessageFunction<K extends MessageID>(
  * because no one is listening, listen for `<messageId>${ READY_MESSAGE_SUFFIX }` messages for 6 seconds, and try to send the message again
  * once that's received, or throw a time out error otherwise.
  */
-export async function iframeIsomorphicSendMessage<K extends MessageID>(
+export async function isomorphicSendMessage<K extends MessageID>(
   messageData: MessageData<K>
 ) {
   // See the "Receive API calls" comment in `ArConnect/src/contents/api.ts` for more on message passing.
@@ -166,7 +166,7 @@ export async function iframeIsomorphicSendMessage<K extends MessageID>(
   });
 }
 
-export function iframeIsomorphicOnMessage<K extends MessageID>(
+export function isomorphicOnMessage<K extends MessageID>(
   messageId: K,
   callback: OnMessageCallback<K>
 ): void {
